apply fixes from generate-snippets branch

+++ ApiDoctor.Console/Program.cs
@@ -711,28 +711,6 @@ namespace ApiDoctor.ConsoleApp
             return results;
         }
 
-        private static DocFile[] GetSelectedFiles(BasicCheckOptions options, DocSet docset)
-        {
-            List<DocFile> files = new List<DocFile>();
-            if (!string.IsNullOrEmpty(options.FilesChangedFromOriginalBranch))
-            {
-                GitHelper helper = new GitHelper(options.GitExecutablePath, options.DocumentationSetPath);
-                var changedFiles = helper.FilesChangedFromBranch(options.FilesChangedFromOriginalBranch);
-
-                foreach (var filePath in changedFiles)
-                {
-                    var file = docset.LookupFileForPath(filePath);
-                    if (null != file)
-                        files.Add(file);
-                }
-            }
-            else
-            {
-                files.AddRange(docset.Files);
-            }
-            return files.ToArray();
-        }
-
         /// <summary>
         /// Parse the command line parameters into a set of methods that match the command line parameters.
         /// </summary>
@@ -1934,16 +1912,6 @@ namespace ApiDoctor.ConsoleApp
             //we are not out to validate the documents in this context.
             options.IgnoreErrors = options.IgnoreWarnings = true;
 
-            var helper = new GitHelper(options.GitExecutablePath, options.DocumentationSetPath);
-            //cleanup any changes that are unstaged/uncommited in repo
-            helper.ResetChanges();
-            helper.CleanupChanges();
-
-            var branchName = options.SourceBranch ?? "snippets-generator-pipeline";
-
-            helper.CheckoutBranch(branchName);
-            FancyConsole.WriteLine(FancyConsole.ConsoleSuccessColor, $"Checking out new branch: {branchName}");
-
             //scan the docset and find the methods present
             var docset = docs ?? await GetDocSetAsync(options, issues);
             if (null == docset)
@@ -1953,7 +1921,6 @@ namespace ApiDoctor.ConsoleApp
             var methods = FindTestMethods(options, docset);
 
             ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12;
-            var languages = options.Language.Split(',');
 
             FancyConsole.WriteLine(FancyConsole.ConsoleSuccessColor, "Generating snippets from Snippets API..");
 
@@ -1964,8 +1931,9 @@ namespace ApiDoctor.ConsoleApp
             WriteHttpSnippetsIntoFile(snippetsPath, methods, issues);
 
             GenerateSnippets(options.SnippetGeneratorPath, // executable path
-                "--SnippetsPath", snippetsPath, "--Languages", options.Language); // args
+                "--SnippetsPath", snippetsPath, "--Languages", options.Languages); // args
 
+            var languages = options.Languages.Split(',');
             foreach (var method in methods)
             {
                 foreach (var lang in languages)
@@ -1995,43 +1963,6 @@ namespace ApiDoctor.ConsoleApp
 
             // clean up
             Directory.Delete(snippetsPath, true /* recursive */);
-
-            if (options.SkipPublishingChanges)
-            {
-                return true;
-            }
-
-            //stage any changes to git 
-            helper.StageAllChanges();
-
-            //check if there is any diff to commit
-            if (helper.ChangesPresent())
-            {
-                FancyConsole.WriteLine(FancyConsole.ConsoleSuccessColor, $"Commiting changes to disk.");
-                var commitMessage = options.CommitMessage ?? "Updated docs by API doctor";
-                helper.CommitChanges(commitMessage);
-
-                //setup for push and pull request
-                GitHub.RepositoryUrl = helper.GetRepositoryUrl();
-                GitHub.AccessToken = options.GithubToken;
-
-                //push changes upstream
-                FancyConsole.WriteLine(FancyConsole.ConsoleSuccessColor, $"Pushing changes upstream");
-                helper.PushToOrigin( GitHub.AccessToken, GitHub.RepositoryUrl, branchName);
-
-                //if target branch is not specified, default to master
-                var targetBranch = options.TargetBranch ?? "master";
-
-                //Create the Pull request
-                FancyConsole.WriteLine(FancyConsole.ConsoleSuccessColor, $"Creating Github Pull Request");
-                var pullRequestTitle = options.PullRequestTitle ?? "Snippet updates by API doctor ";
-                await GitHub.CreatePullRequest(branchName, targetBranch, pullRequestTitle, commitMessage);
-
-            }
-            else
-            {
-                FancyConsole.WriteLine(FancyConsole.ConsoleSuccessColor, "No changes found in docs to commit/push");
-            }
 
             return true;
         }
@@ -2068,20 +1999,11 @@ namespace ApiDoctor.ConsoleApp
             var parser = new HttpParser();
             foreach (var method in methods)
             {
+                HttpRequest request;
                 string snippetPrefix;
                 try
                 {
                     snippetPrefix = GetSnippetPrefix(method);
-                }
-                catch (ArgumentException)
-                {
-                    // we don't want to process snippets that don't belong to a version
-                    continue;
-                }
-
-                HttpRequest request;
-                try
-                {
                     request = parser.ParseHttpRequest(method.Request);
                 }
                 catch (Exception e)
@@ -2096,7 +2018,7 @@ namespace ApiDoctor.ConsoleApp
 
                 var fileName = snippetPrefix + "-httpSnippet";
                 var fileFullPath = Path.Combine(tempDir, fileName);
-                FancyConsole.WriteLine(FancyConsole.ConsoleSuccessColor, $"writing {fileFullPath}");
+                FancyConsole.WriteLine(FancyConsole.ConsoleSuccessColor, $"Writing {fileFullPath}");
 
                 File.WriteAllText(fileFullPath, request.FullHttpText(true));
             }
@@ -2135,9 +2057,9 @@ namespace ApiDoctor.ConsoleApp
                               $"[!INCLUDE [sdk-documentation](../{relativePathFolder}{includeSdkFileName})]\r\n";
 
             const string includeSdkText = "<!-- markdownlint-disable MD041-->\r\n\r\n" +
-                                          "> Read the [SDK documentation](https://docs.microsoft.com/en-us/graph/sdks/sdks-overview) " +
-                                          "for details on how to [add the SDK](https://docs.microsoft.com/en-us/graph/sdks/sdk-installation) to your project and " +
-                                          "[create an authProvider](https://docs.microsoft.com/en-us/graph/sdks/choose-authentication-providers) instance.";
+                                          "> Read the [SDK documentation](https://docs.microsoft.com/graph/sdks/sdks-overview) " +
+                                          "for details on how to [add the SDK](https://docs.microsoft.com/graph/sdks/sdk-installation) to your project and " +
+                                          "[create an authProvider](https://docs.microsoft.com/graph/sdks/choose-authentication-providers) instance.";
 
             /*
                 Scan through the file to find the right line to inject a snippet.
@@ -2224,7 +2146,11 @@ namespace ApiDoctor.ConsoleApp
                     /* DUMP THE SDK LINK FILE */
                     var sdkLinkDirectory = Directory.GetParent(Path.GetDirectoryName(method.SourceFile.FullPath)) + "/" + relativePathFolder;
                     Directory.CreateDirectory(sdkLinkDirectory);
-                    File.WriteAllText(sdkLinkDirectory + "/" + includeSdkFileName, includeSdkText);
+                    // only dump a new file when it does not exist.
+                    if (!File.Exists(sdkLinkDirectory + "/" + includeSdkFileName))
+                    {
+                        File.WriteAllText(sdkLinkDirectory + "/" + includeSdkFileName, includeSdkText);
+                    }
                     break;
                 }
                 case "AdditionalTabInsertion":
@@ -2305,6 +2231,12 @@ namespace ApiDoctor.ConsoleApp
             if (request.Url.Contains("%3E"))
             {
                 request.Url = request.Url.Replace("%3E", "%27");
+            }
+
+            //replace instance " " with single quotes parameter to prevent api fails
+            if (request.Url.Contains(" "))
+            {
+                request.Url = request.Url.Replace(" ", "%20");
             }
 
             return request;
